Extract sample processing helpers
